refactor: rename module to XOpsAgent and reorganize packages
