+++ analyzer/rootcause.go
@@ -1,7 +1,7 @@
 package analyzer
 
 import (
-	"github.com/yourname/ops-agent-poc/model"
+	"github.com/yourname/XOpsAgent/internal/model"
 )
 
 // InferRootCause tries to infer the root cause of a trace.
