+++ go.mod
@@ -1,4 +1,4 @@
-module github.com/yourname/ops-agent-poc
+module github.com/yourname/XOpsAgent
 
 go 1.23.0
 
@@ -57,7 +57,6 @@ require (
 	github.com/prometheus/common v0.65.0 // indirect
 	github.com/prometheus/procfs v0.16.1 // indirect
 	github.com/rogpeppe/go-internal v1.14.1 // indirect
-	github.com/sqlc-dev/pqtype v0.3.0 // indirect
 	github.com/twitchyliquid64/golang-asm v0.15.1 // indirect
 	github.com/ugorji/go/codec v1.3.0 // indirect
 	go.opentelemetry.io/auto/sdk v1.1.0 // indirect
