+++ internal/input/metrics.go
@@ -1,6 +1,6 @@
 package input
 
-import "github.com/yourname/ops-agent-poc/model"
+import "github.com/yourname/XOpsAgent/internal/model"
 
 // FetchMetrics retrieves metrics from a metrics source such as Prometheus.
 func FetchMetrics(source string) []model.Metric {
