+++ internal/model/types.go
